Move error class into flit.common
