+++ flit/common.py
@@ -57,6 +57,15 @@ class NoDocstringError(ProblemInModule): pass
 class NoVersionError(ProblemInModule): pass
 class InvalidVersion(ProblemInModule): pass
 
+class VCSError(Exception):
+    def __init__(self, msg, directory):
+        self.msg = msg
+        self.directory = directory
+
+    def __str__(self):
+        return self.msg + ' ({})'.format(self.directory)
+
+
 @contextmanager
 def _module_load_ctx():
     """Preserve some global state that modules might change at import time.
